// src/types/generated/Task.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskStatus } from "./TaskStatus";
import type { TextFormat } from "./TextFormat";

export type Task = { id: string, storyId: string, parentTaskId: string | null, title: string, description: string, descriptionFormat: TextFormat, result: string, resultFormat: TextFormat, status: TaskStatus, sortOrder: number, createdAt: string, updatedAt: string, deletedAt: string | null, };